# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BookHaven Library - Travel around the world with books.",
        "title": "BookHaven API",
        "contact": {
            "name": "Harya Kumuda",
            "email": "hkkoostanto@students.hacktiv8.ac.id"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/user/topup": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Generates an invoice for top-up using Xendit. The user must be authenticated, and the API request includes the amount to top up. The function handles errors, processes the Xendit response, and updates the transaction history in the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Top Up User Account",
                "parameters": [
                    {
                        "description": "Top Up Request Data",
                        "name": "topUpRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TopUpRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TopUpResponseXenditDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.TopUpErrorResponseXenditDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.TopUpErrorResponseXenditDto": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
        "models.TopUpRequestDto": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                }
            }
        },
        "models.TopUpResponseXenditDto": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "expiry_date": {
                    "type": "string"
                },
                "invoice_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}